requirements.txt
PySide6>=6.6.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0
